✨🚚🧪📝 Unify per-tick kernels into a return-counts menagerie (all u16 timers)
Refactor the transmission and step kernels so they mutate only the per-agent
arrays and RETURN per-node counts; the model applies the deltas to whatever
census it maintains via a new move_count() helper. Timers are now u16 across
the board. This covers all eight SI/SEI/SIS/SEIS/SIR/SEIR/SIRS/SEIRS models with
five kernels.

New kernels (src/rust/src/transmission.rs, steps.rs):
- transmission(...) -> per-node infection counts (S->E or S->I, sets u16 timer);
  transmission_si(...) -> counts (S->I absorbing, SI model).
- step_si / step_sir(absorbing_state = S|R) / step_sirs(imm_dur) -> named lists
  of per-node transition counts (waned/onset/cleared/recovered/waned_r). Each is
  a single u16-timer pass leading with M->S.
- mortality(...) -> list(m,s,e,i,r); births(...) -> list(count, born).
- R helper move_count(from, to, counts, tick) (R/transitions.R).

Removed sir.rs (sir_step, calc_foi moved to transmission.rs) and measles.rs
(measles_step); the old census-writing transmission / transmission_u16. vitals
(constant_pop_vitals_sir, import_infections) updated to u16 timers.

All six examples migrated; attack-fraction pair still validates Kermack-McKendrick
with R0 = beta*D. Tests reworked: test-steps, test-move-count, rewritten
test-transmission / test-mortality / test-births; deleted the obsolete kernel
tests. Docs (CLAUDE.md, READMEs, _pkgdown.yml) updated with the menagerie table.
Full suite 417 pass.

+++ tests/testthat/test-import-infections.R
@@ -20,7 +20,7 @@ test_that("import_infections activates reserved slots and updates census + flow"
   #      at tick 0 is (1, 0); and tick 1's entry is left alone (processed on its tick).
   # Failure would mean imports land in the wrong slot/node/tick or miscount the census.
   state  <- mk("u8",  c(I, I, I, I, S, S))     # slots 4,5 reserved (state S, inactive)
-  timer  <- mk("u8",  c(3, 3, 3, 3, 0, 0))
+  timer  <- mk("u16",  c(3, 3, 3, 3, 0, 0))
   nodeid <- mk("u16", c(0, 0, 1, 1, 0, 0))
   i_count <- allocate_vector("i32", 3L, 2L)
   importations <- allocate_vector("i32", 2L, 2L)
@@ -48,7 +48,7 @@ test_that("import_infections imports nothing on a tick with no schedule entries"
   # When import_infections runs for tick 0
   # Then the count is unchanged, no slot is touched, and the census/flow stay zero.
   # Failure would mean it imports on the wrong tick.
-  state  <- mk("u8",  c(I, S)); timer <- mk("u8", c(2, 0)); nodeid <- mk("u16", c(0, 0))
+  state  <- mk("u8",  c(I, S)); timer <- mk("u16", c(2, 0)); nodeid <- mk("u16", c(0, 0))
   i_count <- allocate_vector("i32", 3L, 1L)
   importations <- allocate_vector("i32", 2L, 1L)
 
@@ -68,7 +68,7 @@ test_that("import_infections aggregates multiple entries for the same tick", {
   # When import_infections runs for tick 0
   # Then 5 agents activate (count 0 -> 5), the I census at tick+1 is (2, 3), and the
   #      flow is (2, 3). Failure would mean per-tick entries are not summed per node.
-  state  <- mk("u8",  rep(S, 5L)); timer <- mk("u8", rep(0, 5L)); nodeid <- mk("u16", rep(0, 5L))
+  state  <- mk("u8",  rep(S, 5L)); timer <- mk("u16", rep(0, 5L)); nodeid <- mk("u16", rep(0, 5L))
   i_count <- allocate_vector("i32", 2L, 2L)
   importations <- allocate_vector("i32", 1L, 2L)
 
@@ -88,7 +88,7 @@ test_that("import_infections errors when imports would exceed capacity", {
   # When import_infections runs
   # Then it errors rather than writing past the allocated buffer.
   # Failure would risk an out-of-bounds slot write.
-  state  <- mk("u8",  c(I, I)); timer <- mk("u8", c(1, 1)); nodeid <- mk("u16", c(0, 0))
+  state  <- mk("u8",  c(I, I)); timer <- mk("u16", c(1, 1)); nodeid <- mk("u16", c(0, 0))
   i_count <- allocate_vector("i32", 2L, 1L)
   importations <- allocate_vector("i32", 1L, 1L)
